Study_2/LLM_based/requirements.txt
# Core model training/inference
torch>=2.1.0
transformers>=4.44.0
accelerate>=0.31.0
peft>=0.11.1
sentencepiece>=0.1.99

# Dataset processing and configs
datasets>=2.20.0
fsspec>=2023.10.0
pyyaml>=6.0.1
